agent_codebase_genius/Server/requirements.txt
jaclang
requests
gitpython
python-dotenv
tree_sitter
tree-sitter-python
openai





